refactor: use-profile-image hook 생성
